Add files via upload
"Final Project" was a group project that I compiled at the end of the year for Data Structures. In it, my partner and I used trees to create a company database for employees, in this case, we used the characters and company from the TV series "The Office". 

Some features it allows you to do is save the information of employees, such as name, position, salary, and experience with "Dunder Mifflin Paper Company". The purpose was to sort employees by salary. The one at the top would become the regional manager